Change `--extra_toolchains` to not accumulate values.
This effectively reverses the current semantics. Previously, the toolchains added via `--extra_toolchains` were all accumulated, in the order the flags were parsed, and added as the highest priority to the list of available toolchains. Unfortunately, this makes it practically impossible to use `--extra_toolchains` to override a toolchain already added by `--extra_toolchains`, especially if it was added in a blazerc file.

With this change, only the _last_ `--extra_toolchains` will have any effect, but multiple comma-separated toolchains can be added. The toolchains will still be kept in order defined in the single value used.

Fixes #19945.

RELNOTES: The `--extra_toolchains` flag no longer accumulates values, instead the last usage of `--extra_toolchains` is used and others are ignored.
PiperOrigin-RevId: 576922897
Change-Id: I80261ad0c63735a8dcd42e17d948c842dca700c5