Fix crash when looking up tree metadata while prefetching the output of an action template expansion.
An action resulting from an action template expansion produces only one of the files in a tree artifact. If, in addition, the file is an in-memory output but its download was requested by the --remote_download_* flags, it will be prefetched when finalizing the action, at which point the prefetcher will look at the metadata for the full tree to determine whether the file should be materialized as a symlink into a different location. This results in a crash because the full tree metadata isn't yet available.

As a fix, the prefetcher now tolerates the absence of tree metadata when these conditions apply. This is still correct because an action template expansion will never materialize its outputs as symlinks.

Fixes #19988.

Closes #20001.

PiperOrigin-RevId: 578176873
Change-Id: I64f765ca5f8852d86150ad866a5fc5930092d661