Update tracking of rewound keys in rewinding tests to prepare for incremental builds.
The listener that collects "dirtied" keys is modified to tolerate and ignore `DirtyType` values besides `DirtyType#REWIND`, since when tests with incremental builds are added, there are calls with other `DirtyType` values that aren't related to rewinding.

Most of the changes are from renaming "dirtied" to "rewound" to better reflect what is being tracked. The only change in logic is in `RewindingTestsHelper#collectOrderedRewoundKeysListener`.

PiperOrigin-RevId: 579179456
Change-Id: I1dae384cc5bcfddb06f0a367a72b626be3698cce