Rollforward of https://github.com/bazelbuild/bazel/commit/f7946d0107dd75b2f45bcc79b91c016d075a756d: Propagate graph_node_aspect over a computed default attribute
Set computed default to deps only when we need to propagate graph_node aspect.

This change makes it possible to collapse cc_binary and cc_test into a single rule.

NEW: using initializers setting a private field (instead of computed default that caused a regression). This is a workaround that propagates an aspect optionally. Allowing initializers set a private field is a potential replacement for computed default. Current implementation of initializers allows this only in builtins - we need to be careful, because implicit dependencies potentially affect other system.
PiperOrigin-RevId: 580118899
Change-Id: I4625c0c4f903b19e697f8d5ae09bbea682d78a13