Start adding async deserialization APIs. Make DynamicCodec use it.
There are a few changes needed for async serialization. This
introduces the first: not requiring DeserializationContext.deserialize
to immediately return the value. This is achieved by adding two
new deserialize overloads, one taking an (object, offset) input and
one taking a FieldSetter callback. Using these methods allows the
DeserializationContext to call them at a later point in time.

This change refactors DynamicCodec to use the new overloads, making
it fully asynchronous compatible. Instead of using a large top-level
if-else type switch, DynamicCodec's main control flow is applied using
polymorphism, which is more efficient. The primary motivation is to allow
the type structure to better reflect the logic. FieldHandler could have
been implemented by the previous TypeAndOffset class, but then it would
be implemented in many places where it would not be appropriate to invoke.

This restructuring necessitates a new array recursion approach. Arrays
now have their own custom handler. One upside is that the inner loop of
an array serialization becomes much tighter and no longer includes the
outer if-else type switch per-element.

PiperOrigin-RevId: 580275120
Change-Id: I4c3bc132a46609ec67641ca4cc89c7205ba83249