Separate main aspect prerequisites from rule and base aspects prerequisites
add `separate_aspect_deps` flag to guard this fix, when the flag is enabled the dependencies of the main aspect will be separated from those of the target and the base aspects. `ctx.attr.{attr_name}` will always get the attribute value from the main aspect and `ctx.rule.attr.{attr_name}` will get the value from the rule if it has an attribute with that name or from the base aspects attributes (first one in the aspects path wins).

PiperOrigin-RevId: 580292823
Change-Id: I719f8775c0c2aa582aed600387cb6b321cf5b639