[7.0.0] Delete the now redundant --experimental_execution_log_file flag, and optimize creation of unsorted logs. (#20132)
Since we're about to introduce a new log format (see
https://github.com/bazelbuild/bazel/issues/18643), let's take this
opportunity to clean up the command line API.

Avoid a pointless copy of the execution log, if no sorting is requested,
by writing directly into the requested output binary path. If a
temporary file is necessary, write it to a well-known path in the output
base, and delete it after conversions are successful; this ensures that
large temporary files don't accumulate in /tmp.

RELNOTES[INC]: Delete the --experimental_execution_log_file flag. Use
--execution_log_binary_file in conjunction with --noexecution_log_sort
instead.

PiperOrigin-RevId: 580132026
Change-Id: Icca827260176b3ea804384d2bed4a56ce85cfc14