[7.0.0] [Skymeld] Don't plant the symlinks to the paths specified in  (#20168)
Bazel in skymeld mode eagerly plants the symlinks from the execroot to
every entries under the source tree. This leads to some pain points for
certain groups of users
(https://github.com/bazelbuild/bazel/issues/19581).

This CL offers a workaround for that issue. The users can now list the
paths that they don't want Bazel to symlink to in the `.bazelignore`
file in their project.

Commit
https://github.com/bazelbuild/bazel/commit/caf1702ce9fa060169548181afb9e19956826f0b

PiperOrigin-RevId: 581919915
Change-Id: I329365b0655fc0bcb2db1eeea91ef74c775c72ae

Co-authored-by: Googler <leba@google.com>