[7.0.0] Test path mapping in Bazel (#20178)
Add end-to-end tests to verify that `--experimental_output_paths=strip`
works with sandboxed, sandboxed worker, and remote execution and results
in the desired cache hits. Also adapt and enable
`config_stripped_outputs_test` for Bazel (e.g. make it work on macOS).

This requires a minor changes: We ensure that full `.jdeps` are
rewritten to contain unmapped paths even after a fallback to full
classpath.

Closes #18155.

Commit
https://github.com/bazelbuild/bazel/commit/cab7c6be21918410cda85a8f48eeb7840e3855c7

PiperOrigin-RevId: 582014351
Change-Id: If9c51acb090d9fc4bc86052143ad9960e44fd8a7

Co-authored-by: Googler <gregce@google.com>