Copy and lift labels in arguments of initializers
The arguments are copied into immutable Starlark values. They are type checked.  Strings are converted to labels (in the right context).

I was worried copying would cause memory regression, but it doesn't. (There's initializer on cc_binary).

There's a potential optimisation to use Starlark types everywhere. StarlarkList, Dict are already as good as ImmutableList and ImmutableMap (I think). We'd need to converge Selectors to a single type. Then the convert operation becomes the same as copyAndLiftStarlarkValue. And it could be optimised to a no-op when called twice.

PiperOrigin-RevId: 582324735
Change-Id: I64ee4f70fa12f9818d57474ba33d085aeed347ad