Add a native image of turbine to the prebuilt Java tools
A simple local experiment shows that the time for all Java header compilation actions required for `//src:bazel-dev` decreases by a factor of 4.5 when using a native image of turbine instead of jar. The time taken for an incremental build of `//src/main/java/com/google/devtools/build/lib/bazel:BazelServer` after adding a public method to `Label` decreases by a factor of 2 with `--experimental_java_classpath=bazel`.

As a first step towards using the native image in Java toolchains, ship it as part of the prebuilt Java tools by using rules_graalvm.

Closes #19361.

PiperOrigin-RevId: 582531539
Change-Id: If71d8ec86fecbc1363a748d197f88962bd489fb4