Deprecate FilesToRunProvider.getFilesToRun().
It has very confusing semantics: it represents the filesToBuild of the target in question (which is not necessarily the same as its executable, but usually is), its runfiles middleman (which is not the runfiles tree -- building it materializes the runfiles tree on local disk and the artifacts in it, but does not add either to the inputs of an action with the middleman on its inputs) and the executable, but only for Starlark rules (wat)

Alternatives are in the Javadoc added.

RELNOTES: None.
PiperOrigin-RevId: 582561563
Change-Id: Ie00a7f1fd0b8e271aa119b9f1303b297684e8882