Get the genrule setup script using FileProvider instead of FilesToRunProvider.
What the latter does is not very well defined. It's much simpler conceptually to say "the files this dependency builds".

This approach is dead simple and works as long as the setup script is just a file, which is the case both in Bazel and at Google. If it was a full binary, we'd have to worry about its runfiles, but that's not the case and it doesn't work at HEAD anyway because its hypothetical runfiles are not added to a RunfilesSupplier.

RELNOTES: None.
PiperOrigin-RevId: 582561869
Change-Id: Iad507c0544c5879ebc3eacd793d1885c5136a22f