Remove Runfiles.Builder.addLegacyExtraMiddleman().
Its last call site was in android_device. My best guess from the history of this line of code (see https://cr.bazel.build/31030 and issue #4090) is that its intended effect was that the runfiles tree of the unified Android launcher (?) should be built when the android_device rule is.

However, I have evidence that it's not necessary: if a particular runfiles tree is needed for the execution of an action, it should be part of the Spawn representing that action. The fact that the tests pass seem to imply that this is already the case.

A weak additional piece of evidence is that the runfiles middleman is already in extraFilesToRun, which in turn becomes parts of the FilesToRun of android_device.

RELNOTES: None.
PiperOrigin-RevId: 582572750
Change-Id: Id70e01501b3de7665941703efb7c457d7a03d434