Do not add the runfiles middleman of the unified launcher to the files to run of android_device .
This makes it so that when one builds the android_device rule (or something that calls getFilesToRun() on it), the runfiles tree of the unified launcher will not get built (however, the runfiles of the unified launcher will still get merged into the runfiles of the android_device rule)

Evidence that this is the right thing to do is (in addition to the passing tests) is that the prevalent pattern in Android rules is to use Runfiles.Builder.merge(RunfilesSupport), which does the same thing and is exclusively used in android_* rules.

If one wanted to have a separate runfiles tree in an action, one would need to do two additional things:

- Put the middleman on the inputs of the action in question (this causes the files in the runfiles tree be built and creates the runfiles tree locally)
- Put the desired runfiles into the RunfilesSupplier of the action (this is never done in the Android rules)

So the only case that could conceivably be broken is when local actions are involved.

RELNOTES: None.
PiperOrigin-RevId: 582966126
Change-Id: I1717b1c53ded1476c4c5e2dad95fa4cbdcc197f6