Remove the ability to add runfiles middlemen to actions creating deploy jars.
This was only used in android_local_test and seems to be completely wrong: the deploy jar does not need the runfiles of the binary to build the deploy jar and even if it needed them, this alone would not put them on the inputs of the action: to do that, it's not enough to depend on the runfiles middleman, it's also necessary for the action to have an appropriate RunfilesSupplier, which it didn't.

(the code removed here would make the runfiles tree available for local actions, though, but not remote ones)

RELNOTES: None.
PiperOrigin-RevId: 582976603
Change-Id: Ide9c735a725c5eb25d81658332b18bca373d35fc