Add --incompatible_java_info_merge_runtime_module_flags
Change the behavior of `JavaInfo` constructor to be more like `java_info_for_compilation`, by merging in the `add_exports` and `add_opens` flags for `runtime_deps` in addition to just `deps` and `exports`. Guard it under an `--incompatible_` flag which defaults to false, but I'm hoping to make it default to true in 8.x.

Second half of #20033

Closes #20037.

PiperOrigin-RevId: 582982387
Change-Id: Ibff680f71efed82f20da7d9ee83f0bfa7e5f5697