Simplify the specification of `NodeEntry#isDirty` and update implementations.
Previously, the implementation of `AbstractInMemoryNodeEntry#isDirty` was:

> `return !isDone() && dirtyBuildingState != null;`

The method definition said:

> Returns true if the entry is new or marked as dirty.

The two were inconsistent for a newly created entry (not done, null `dirtyBuildingState`), although it didn't cause issues because `isDirty()` was only being called on a node in the newly created state in one place (`toValue()`, but no longer after this change).

For clarity, define `isDirty()` as the inverse of `isDone()`, meaning that a newly created node returns `true`.

PiperOrigin-RevId: 583168878
Change-Id: Ideb42b88a369ed1f9a630c7d7bfb8099e3dd5290