[7.0.0] Emit a fixed config segment with path stripping (#20236)
Instead of outright removing the config segment of exec paths with the
stripping path mapper, replace it with a fixed string (`cfg`). This has
the following benefits:

* Actions that make assumptions about the general exec path structure
(e.g. because they use the Bash runfiles library, which parses the exec
path and expects a `bazel-out/<something>/bin/...` format) continue to
work.
* Path stripping becomes idempotent, which fixes issues in case an
artifact's path is mapped both structurally and via the string-based
stripping.

Closes #20219.

Commit
https://github.com/bazelbuild/bazel/commit/0d0caba2e940f9eec2051c43c86954cf9c95635f

PiperOrigin-RevId: 583266215
Change-Id: I063fd7fd502a92660952427c5ce893eacd16436d

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>