Rename the global attribute "applicable_licenses" to "package_metadata" but allow both spellings.
Apology to the world. The name "package_metadata" is not great. No name
is great. We can't just use "metadata" - that will certainly clash with
existing rules. We could call it "applicable_metadata", but that makes
no historic sense. The older name as "applicable_licenses" because
we needed a name that was not just "licenses".  "target_metadata"
might be better, but, the path of least confusion is to pair it with
"default_package_metadata" and call it "package_metadata". #namingishard

Phase 2: Make the change

RELNOTES: The global "applicable_licenses" attribute has been renamed to "package_metadata". The older name may be used for creating targets, but starlark code must use the name "package_metadata" to retrieve it.  Breaking change.
PiperOrigin-RevId: 583397119
Change-Id: I21473b0730b6c70111be84d10b927d370bec0462