Offset the bytes attributed to jdk.internal.vm.FillerArray from the retained heap metrics.
With JDK21, an arbitrary amount of FillerArray instances is retained on the heap at the end of each command, which fluctuates our heap metrics a lot.

This CL adds a method that looks into the heap histogram and gets the total #bytes attributed to the FillerArray. We'll offset that against the final value to mitigate the instability.

PiperOrigin-RevId: 583582766
Change-Id: I7f7b7b993b76a655952baab1e92f27919e31f8f6