Apply the heap size offset to `info used-heap-size-after-gc` too.
This is a follow up of https://github.com/bazelbuild/bazel/commit/6ef79e15496db51961a4ed94703d65227480b845.

We're only applying this to the places where the _retained_ heap is measured, and not _peak_ heap. Peak heap is A) quite unstable anyway and B) measured multiple times during a build, which means the cost of constructing the histogram might add up.

PiperOrigin-RevId: 584018533
Change-Id: Ic48ce5af84fb2cf4713ae261d30e5cc308883080