[7.0.0] Build failures and internal errors when switching from 7.0.0rc3 to 7.0.0rc4 (#20278)
The methods are documented as such in FileSystem. If we don't do this,
there will be a discrepancy between getFastDigest and stat, as the
latter can follow symlinks. This can manifest as a crash (see #20246) as
the digest computation will take the missing fast digest for a symlink
as a signal to compute the digest manually; this would fail when the
symlink target is an in-memory file, which doesn't have an associated
inode as required to compute the cache key (see
DigestUtils#manuallyComputeDigest).

Fixes #20246.

Commit
https://github.com/bazelbuild/bazel/commit/aab19f75cd383c4b09a6ae720f9fa436bf89d271

PiperOrigin-RevId: 584297990
Change-Id: I65e586ea84635a279208e24c421f54ae46ee21b8

Co-authored-by: Googler <tjgq@google.com>