Pass `ct.sym` to direct version of Turbine
This ensures that Turbine has access to `ct.sym` and thus supports the
`--release` flag even if it is compiled into a Graal native image, which
doesn't have access to a `JAVA_HOME`.

This requires exposing the `ct.sym` file on a new `lib_ct_sym` attribute
of `java_runtime` and passing it into Turbine via the
`turbine.ctSymPath` system property introduced in v0.3.1.