Add `CommandExtensionReporter` to `CommandEnvironment` to enable streaming back command extension messages.
The `RunResponse.command_extensions` field is currently populated at the end of
a command. Since `Run` is a streaming operation we can conceivably stream back
command extensions if the build tool were given an appropriate interface to do
so.

The normal Bazel client (the C++ binary) will currently drop these extensions on the floor so users will need to connect directly to the Bazel server to retain the
structured metadata returned in command extensions. A possible enhancement would extend the C++ client to optionally write response extensions to a file or files.

PiperOrigin-RevId: 584426022
Change-Id: Ie0c894d09ac8b2ddd1b1419396ac292a61b1af77