[JSON Trace Profile] Correctly identify GC thread
Currently, the thread that includes the garbace colleciton notification
events is not correctly identified.
This means that in the JSON trace profile, the thread:

* is not renamed, as desired.
* is not sorted towards the top, as desired.

The thread is identified by its name, which has changed from "Service Thread" to
"Notification Thread". This change updates the profile code accordingly.

In response to https://github.com/bazelbuild/bazel/issues/18548#issuecomment-1823361627