[JSON Trace Profile] Add `tid` to counter series
The JSON trace profile includes counter for data such as
action count, CPU usage, memory usage.
The code for writing the events to include these was refactored in
https://github.com/bazelbuild/bazel/commit/a03674e6297ed5f6f740889cba8780d7c4ffe05c
As part of that change the events written no longer include a
`tid` (thread ID). However, each event in the JSON trace profile is expected to
include a `tid`, as documented in
https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview#heading=h.uxpopqvbjezh

This change ensures the counter series events again have a `tid`, using
the current thread's ID. This is consistent with the behavior that was present before
the refactor.

In response to https://github.com/bazelbuild/bazel/issues/18548#issuecomment-1823361627