[7.0.0] Consider each runfiles tree to be a distinct node when walking the spawn inputs.
This is a similar idea to https://github.com/bazelbuild/bazel/commit/0f55d124c05ee2654c3ff56e0332e71ff1eb437c, where we did the same for tree artifacts.

Currently, the sole beneficiary of this change is the Merkle tree cache, which becomes more efficient in the presence of runfiles trees shared by multiple actions. In the future, we might reuse this code path to produce a more compact execution log.

PiperOrigin-RevId: 585582119
Change-Id: I2caec62eb26cab6dc2530889ef075acc1d9e93d9