Expose `cc_helper.tokenize` and optionally use it in `java_helper.tokenize_javacopts`
For large option sets, such as those involved in java compilation, `ctx.tokenize()` is up to 100x faster than the pure Starlark implementation. We should resort to using the slow path only in the cases where `ctx` is unavailable (such as when performing a `map_each` in `args.add_all`).

PiperOrigin-RevId: 585977311
Change-Id: I547c6dbc1a37ef15494a9b398a597a4246c08ea9