Add missing googletest directory to cpp usecase doc to reflect actual structure
The use-case doc assumes that `googletest`'s directories such as `src`/`include` are located at the top, but if we see the contents of https://github.com/google/googletest/archive/release-1.10.0.zip, there is an intermediate directory `googletest`.

![Screenshot 2023-09-26 at 0 05 51](https://github.com/bazelbuild/bazel/assets/1391330/02d56fdb-272c-4a25-ad9c-2f36dece2c1c)

So if we want to have `strip_prefix = "googletest-release-1.10.0"` with the custom `BUILD` file, we need to append `googletest` as a prefix (just like the repository's [`BUILD` file](https://github.com/google/googletest/blob/release-1.10.0/BUILD.bazel)). And so fixing the doc.

Closes #19619.

PiperOrigin-RevId: 586034829
Change-Id: Ie7e1d458d65e47584e1438f99fb1dcb198fbe217