Fix execution platform in the extra action
Before extra action selected shell based on its own execution plaform and was executed on the platform of the action it shadows.
After the change extra action selects shell based on the rule's default execution platform and is executed on the same platform.

This obscured a double problem: it works in integration tests with C++ toolchain resolution disabled, because execution platform is always Linux, but requires-darwin triggers on objc compilation on Mac and not the extra action.
It also works in the depot, because we're using Linux crosstool to compile objc and execute extra action.

However extra actions never worked on Mac with C++ toolchain resolution enabled. Wrong Sh was chosen when they are executed on Mac.

PiperOrigin-RevId: 586369094
Change-Id: I58d3d3df041e358f23767f97d55dcb2211437ac7