Fix `upload_all_java_tools.sh`
The initial bazel invocations to produce the zip files and the later test invocation(s) may not have the exact same flags and can end up with different configurations (and hence output dirs). If that happens, `bazel-bin` points somewhere else, and the tests don't find them.

So we copy the files out of `bazel-bin` before passing them to the test(s).

Tested in https://buildkite.com/bazel-trusted/java-tools-binaries-java/builds/241

PiperOrigin-RevId: 586414889
Change-Id: Iaef1ea0730d937a73f0dac6c9792429955bbe9db