Cache dex syntethic context in dex builder
There is a data race when using workers when there are multiple
libraries that include the same class. This is more likely to happen
when building multiple android_binary targets in a single invocation.
When a cached entry is found, only the dex content is used and the
sythetic info is ignored. In those instances, the dex archive will
be incorrectly created and might lead to failures during dex merging