Isolate the java rules unit tests from the host platform
Parameterizes the unit tests on target platform, ensuring we actually exercise platform specific code paths. Before this, we were (indirectly) relying on the host platform, which in unit tests is hardcoded to linux/x86_64.

This found a bug in the change to `getJvmFlagsForJavaBinaryExecutable` in unknown commit - we should be checking the platform of the `java_{binary,test}` target, not that of the action that creates the stub script / executable.

PiperOrigin-RevId: 587643843
Change-Id: Idcd512e25d3c96cb74ccda836817121cfa5d08a6