[7.0.0] Allow most forms of symlink inside a remotely produced tree artifact, matching a locally produced one. (#20426)
This CL harmonizes the criteria for permitting a symlink for locally and
remotely produced tree artifacts, namely:

1. A symlink to an absolute path is allowed.
2. A symlink to a relative path is allowed, as long as the target is
inside the tree artifact.
3. The target path must exist (a consequence of tree artifacts
transparently dereferencing symlinks; see also #15454).

as enforced by TreeArtifactValue#visitTree (see
https://cs.opensource.google/bazel/bazel/+/master:src/main/java/com/google/devtools/build/lib/skyframe/TreeArtifactValue.java;l=585;drc=de9d1f59915a36229978d46b78a22c9e5389db92).

The admonition about symlinks potentially compromising hermeticity is
still valid, but there's little gain in making the behavior divergent
between local and remote execution. I don't believe we can go in the
other direction and extend the restriction to cover local execution, as
it would break existing projects.

The --incompatible_remote_disallow_symlink_in_tree_artifact flag is
deleted. It was added at a time when symlink resolution was extremely
unreliable when building without the bytes; the state of symlink
handling has improved a lot since then. I'm deleting the flag entirely
(as opposed to making it a no-op) because it was only introduced in the
Bazel 7 tree and hasn't made it into a stable release yet.

Makes #18284 obsolete.

PiperOrigin-RevId: 587691220
Change-Id: I470a8fc523bdbb7577ad5a564b6b2c0acab17d7d