Consolidate referneces to `--experimental_exec_config`.
Also more strongly verify that the Starlark exec transition is set for this build.

I thought about removing the generic `context` variable in `StarlarkExecTransitionLoadingException`'s
constructor. All uses today fill that out with `--experiemntal_exec_config`. But it seems nice to
keep that open in case we generalize uses with a wider API.

PiperOrigin-RevId: 588074546
Change-Id: I14f169e2e44d482c9acb34968a97dbb008149299