Improve the specification and implementation of `NodeEntry#toValue`.
The only semantic change is that a `NonIncrementalInMemoryNodeEntry` that is rewound stores its previous value so that `toValue()` can return it while it is rewinding. This is consistent with how an `IncrementalInMemoryNodeEntry` already works, and since the implementation is now the same, `toValue()` is moved to the shared `AbstractInMemoryNodeEntry`.

The motivation for this is b/314117358: currently, `ArtifactNestedSetFunction` maintains a big map of built artifacts that is mostly redundant with data in the Skyframe graph. One exception is that currently artifacts in the big map are not evicted when their generating action is rewound, so a simultaneously executing action can still read them. Allowing `toValue()` to return a non-null value even for a rewound node makes it work consistently with how the big artifact map is currently used.

An alternative is to treat an unexpectedly-null value as a "lost input" under the assumption that it was rewound, and do more rewinding to recover from that. This may be more principled, however it is much more difficult to implement (we'd need to introduce the possibility of lost inputs before even attempting to executing the action) and is also inferior in the case of an action that generates multiple outputs: if only one output is lost, the entire action is rewound, although another output may still be available. In this case, the alternative approach could result in more rewinding than necessary.

Assertions for `toValue()` are peppered around various unit tests, and one new unit test case is added for the possibility of a node that is rewound and then reset.

PiperOrigin-RevId: 588122054
Change-Id: Ic3f4d7acff85fa626f13430658b7b832763955b1