Generate .pdb files when building cc_shared_library targets for Windows
This enables the functionality to generate .pdb files when building a DLL from a cc_shared_library target with the Lexan toolchain, which is already supported when using a cc_binary target.

The change has been tested manually with a local cc_shared_library target, with and without the shared_lib_name attribute.

I'm not exactly sure how the automated/unit tests for cc_shared_library work, but it seems like when the test targets in //third_party/bazel/src/main/starlark/tests/builtins_bzl/cc/cc_shared_library/test_cc_shared_library/BUILD.builtin_test are built for Windows, the pdb logic should also be executed, so if there are failures they should be caught. A test that would make sense is to check the filenames for the generated pdb files, but I don't see tests that do this for DLLs so I assume we're not doing that for Windows.

PiperOrigin-RevId: 588441321
Change-Id: I20aa6d855514d3864c7f2f4b658f279367128030