Replace the big map in `ArtifactNestedSetFunction` with direct Skyframe graph access.
This is fundamentally equivalent, but saves memory (~0.4% on an example build) from the `ConcurrentHashMap` structure, though all of the keys and values are typically still retained in Skyframe nodes.

Memory savings may be greater on certain incremental builds due to Skyframe GC of dirty nodes (by default, we delete all dirty nodes at the end of a build, see `--version_window_for_dirty_node_gc`). Previously, the associated metadata would not be evicted from the big map in this case.

PiperOrigin-RevId: 588458502
Change-Id: I8e334953fb87741e84b0233c547c4271b8e5c105