Automated rollback of commit ffa75bf996ce52067fae8f530921a1db16fb46b2.
*** Reason for rollback ***

Does not handle the case where a flaky action is rewound and evaluates to an error (toValue returns null). This situation was observed in b/315301248.

*** Original change description ***

Replace the big map in `ArtifactNestedSetFunction` with direct Skyframe graph access.

This is fundamentally equivalent, but saves memory (~0.4% on an example build) from the `ConcurrentHashMap` structure, though all of the keys and values are typically still retained in Skyframe nodes.

Memory savings may be greater on certain incremental builds due to Skyframe GC of dirty nodes (by default, we delete all dirty nodes at the end of a build, see `--version_window_for_dirty_node_gc`). Previously,...

***

PiperOrigin-RevId: 589123293
Change-Id: I71e20d93c4d2d0c2c384cb9e7e9edbb0203d69d0