Don't rewind nodes that are done with an error.
We don't intentionally attempt to rewind nodes in error, but it can happen when two parents observe the same lost input and there is a race to rewind the node. After the first parent rewinds it, the node completes with an error (i.e. because the action is flaky), then the losing parent arrives. Rewinding the errorful node again is problematic in a `--nokeep_going` build because error bubbling expects the errorful node to be done.

A regression test is added to `MemoizingEvaluatorTest`.

PiperOrigin-RevId: 589132267
Change-Id: Idc9c30c2fea34d02d31fa20a5f4855cb1300c1ba